figuring stuff out
still receiving an error in the feedforward function, still can't get to grips with the guy in the book that uses the center of the screen for training but the force pointing to multiple targets for the feedforward? I only need to follow one target so the array thing is confusing me, obfuscating the actual simple process I should be able to see?